# api/openapi.yaml
openapi: 3.1.0
info:
  title: KubeVirt Shepherd API
  description: |
    API for managing KubeVirt virtual machines and cloud-native governance.
    Adheres to ADR-0021: API Contract-First Development.
  version: 1.0.0
  contact:
    name: KubeVirt Shepherd Team
    url: https://github.com/kv-shepherd/shepherd
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: Production API
...
paths:
...
  /admin/users:
    get:
      tags: [admin]
      summary: List users
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
    post:
      tags: [admin]
      summary: Create a local JWT user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/users/{user_id}:
    patch:
      tags: [admin]
      summary: Update a local JWT user
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [admin]
      summary: Delete a local JWT user
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserID'
      responses:
        '204':
          description: User deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{user_id}/role-bindings:
    get:
      tags: [rbac, admin]
      summary: List global role bindings for a user
      operationId: listUserRoleBindings
      parameters:
        - $ref: '#/components/parameters/UserID'
      responses:
        '200':
          description: Role binding list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalRoleBindingList'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [rbac, admin]
      summary: Create global role binding for a user
      operationId: createUserRoleBinding
      parameters:
        - $ref: '#/components/parameters/UserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalRoleBindingCreateRequest'
      responses:
        '201':
          description: Role binding created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalRoleBinding'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/users/{user_id}/role-bindings/{binding_id}:
    delete:
      tags: [rbac, admin]
      summary: Delete a user's global role binding
      operationId: deleteUserRoleBinding
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/RoleBindingID'
      responses:
        '204':
          description: Role binding deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable Parameters (ADR-0023 Pagination) ──────
  parameters:
    Page:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      description: Case-insensitive search text
      schema:
        type: string
...
    UserID:
      name: user_id
      in: path
      required: true
      schema:
        type: string
...
    RoleBindingID:
      name: binding_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: Machine-readable error code (frontend handles i18n)
        message:
          type: string
          description: English log message (not for UI display)
        params:
          type: object
          additionalProperties: true
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # ── System ──────────────────────────────────────
...
    User:
      type: object
      required: [id, username, enabled, created_at]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        display_name:
          type: string
        enabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    UserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
    UserCreateRequest:
      type: object
      required: [username, password]
      properties:
        username:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
        password:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
          format: password
        email:
          type: string
        display_name:
          type: string
        enabled:
          type: boolean
        force_password_change:
          type: boolean
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
        display_name:
          type: string
        enabled:
          type: boolean
        password:
          type: string
          format: password
        force_password_change:
          type: boolean
...
    GlobalRoleBinding:
      type: object
      required: [id, user_id, role_id, role_name, scope_type]
      properties:
        id:
          type: string
        user_id:
          type: string
        role_id:
          type: string
        role_name:
          type: string
        role_display_name:
          type: string
        scope_type:
          type: string
        scope_id:
          type: string
        scope_display_name:
          type: string
        allowed_environments:
          type: array
          items:
            type: string
            enum: [test, prod]
        managed:
          type: boolean
        managed_source:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
    GlobalRoleBindingList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GlobalRoleBinding'
        pagination:
          $ref: '#/components/schemas/Pagination'
    GlobalRoleBindingCreateRequest:
      type: object
      required: [role_id]
      properties:
        role_id:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
        scope_type:
          type: string
          default: global
        scope_id:
          type: string
        allowed_environments:
          type: array
          items:
            type: string
            enum: [test, prod]
...
security:
  - BearerAuth: []